+++ src/main/java/com/github/spotbugs/internal/SpotBugsReportsImpl.java
@@ -2,7 +2,6 @@ package com.github.spotbugs.internal;
 
 import org.gradle.api.Task;
 import org.gradle.api.reporting.SingleFileReport;
-import org.gradle.api.reporting.internal.CustomizableHtmlReportImpl;
 import org.gradle.api.reporting.internal.TaskGeneratedSingleFileReport;
 import org.gradle.api.reporting.internal.TaskReportContainer;
 
@@ -11,11 +10,11 @@ import com.github.spotbugs.internal.spotbugs.SpotBugsXmlReportImpl;
 
 public class SpotBugsReportsImpl extends TaskReportContainer<SingleFileReport> implements SpotBugsReportsInternal {
 
-  public SpotBugsReportsImpl(Task task) {
+public SpotBugsReportsImpl(Task task) {
       super(SingleFileReport.class, task);
 
       add(SpotBugsXmlReportImpl.class, "xml", task);
-      add(CustomizableHtmlReportImpl.class, "html", task);
+      add(SpotBugsHtmlReportImpl.class, "html", task);
       add(TaskGeneratedSingleFileReport.class, "text", task);
       add(TaskGeneratedSingleFileReport.class, "emacs", task);
   }